requirements.txt
absl-py==0.14.0
aiohttp==3.7.4.post0
antlr4-python3-runtime==4.8
appnope==0.1.2
argcomplete==1.12.3
argon2-cffi==21.1.0
async-timeout==3.0.1
attrs==21.2.0
av==8.0.3
backcall==0.2.0
bleach==4.1.0
cachetools==4.2.2
certifi==2021.5.30
cffi==1.14.6
chardet==4.0.0
charset-normalizer==2.0.6
colorama==0.4.4
cycler==0.10.0
debugpy==1.4.3
decorator==5.1.0
defusedxml==0.7.1
entrypoints==0.3
fsspec==2021.9.0
future==0.18.2
google-auth-oauthlib==0.4.6
google-auth==1.35.0
grpcio==1.40.0
h3ds==0.1.1
hydra-core==1.1.1
idna==3.2
imageio==2.9.0
importlib-metadata==4.8.1
importlib-resources==5.2.2
ipykernel==6.4.1
ipython-genutils==0.2.0
ipython==7.28.0
ipywidgets==7.6.5
jedi==0.18.0
jinja2==3.0.1
joblib==1.0.1
jsonschema==3.2.0
jupyter-client==7.0.3
jupyter-console==6.4.0
jupyter-core==4.8.1
jupyter==1.0.0
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.2
kiwisolver==1.3.2
kornia==0.5.11
lpips==0.1.4
markdown==3.3.4
markupsafe==2.0.1
matplotlib-inline==0.1.3
matplotlib==3.4.3
mistune==0.8.4
multidict==5.1.0
nbclient==0.5.4
nbconvert==6.2.0
nbformat==5.1.3
nest-asyncio==1.5.1
notebook==6.4.4
numpy==1.21.2
oauthlib==3.1.1
omegaconf==2.1.1
opencv-python==4.5.3.56
packaging==21.0
pandas==1.1.5
pandocfilters==1.5.0
parso==0.8.2
pexpect==4.8.0
pickleshare==0.7.5
pillow==8.3.2
plotly==5.3.1
prometheus-client==0.11.0
prompt-toolkit==3.0.20
protobuf==3.18.0
ptyprocess==0.7.0
py==1.10.0
pyasn1-modules==0.2.8
pyasn1==0.4.8
pycparser==2.20
pydeprecate==0.3.1
pygments==2.10.0
pyparsing==2.4.7
pyrsistent==0.18.0
python-dateutil==2.8.2
pytorch-lightning==1.4.8
pytz==2021.1
pyyaml==5.4.1
pyzmq==22.3.0
qtconsole==5.1.1
qtpy==1.11.2
requests-oauthlib==1.3.0
requests==2.26.0
rsa==4.7.2
scikit-learn==1.0
scipy==1.7.1
send2trash==1.8.0
six==1.16.0
smmap==4.0.0
tabulate==0.8.9
tenacity==8.0.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorboard==2.6.0
terminado==0.12.1
test-tube==0.7.5
testpath==0.5.0
threadpoolctl==2.2.0
timm==0.4.12
toml==0.10.2
torch==1.9.1
torchmetrics==0.5.1
torchvision==0.10.1
tornado==6.1
tqdm==4.62.3
traitlets==5.1.0
trimesh==3.9.32
typing-extensions==3.10.0.2
urllib3==1.26.7
wcwidth==0.2.5
webencodings==0.5.1
werkzeug==2.0.1
widgetsnbextension==3.5.1
yarl==1.6.3
zipp==3.5.0